// src/app/modules/newsData/newsData.interface.ts
export interface INewsData {
  article_id: string;
  link: string;
  title: string;
  description: string;
  content: string;
  keywords?: string[];
  creator?: string[];
  language?: string;
  country?: string[];
  category?: string[];
  datatype?: string;
  pubDate: string;
  pubDateTZ: string;
  fetched_at: string;
  image_url?: string;
  video_url?: string;
  source_id: string;
  source_name: string;
  source_priority: number;
  source_url: string;
  source_icon: string;
  sentiment: string;
  sentiment_stats: string;
  ai_tag: string;
  ai_region: string;
  ai_summary: string;
  duplicate?: string;
}